Twowaypolr.cv.R Update
Changed return values for polr2cv function and checks for equality of
the number of categories between set and sampled subset as well as
default value of “logistic” for a missing x. Added two new functions
for randomly generating a 3-level categorical response from the
proportional odds and multinomial logit models given a dataset x.